The bundle.js file was not up-to-date which was causing the bugs.

+++ js/bundle.js
@@ -16,12 +16,15 @@ $(() => {
   const socket = window.io();
   //   const $testmes = $('#first-message');
   const $alertUsername = $('.alert-username');
+  const $transferPageHeader = $('.user-name');
   const $alertUsernameBlank = $('.alert-blankusername');
   const $listOfUsers = $('#listOfUsers');
   const $cancelButton = $('#waiting_message .cancel-button button');
+  const fileSendButton = $('#file-send-button');
   const bitrateDiv = document.getElementById('bitrate');
-  const downloadAnchor = document.getElementById('download');
+  const $downloadAnchor = $('#download');
   const statusMessage = document.getElementById('status');
+
   //   const bitrateMax = 0;
   //   const TURN_SERVER_IP = '127.0.0.1';
   let offersForMe = [];
@@ -35,21 +38,12 @@ $(() => {
       },
     ],
   };
-  //   const connection = { optional: [{ DtlsSrtpKeyAgreement: true }
-  //   , { SctpDataChannels: true }] };
-  let myPeerConn; // variable to store the RTCPeerConnection object,
+  let client;
   let ExchangerUsername; // variable for name of requested username
-  let dataChannel;
   //   let offerComplete = false;
   let fileRec;
   let file;
   let sender = false; // to maintain state of the client(whether he/she is a sender or a receiver)
-  let receiveBuffer = [];
-  let receivedSize = 0;
-  let receivedProgress = { value: 0 };
-  let sendProgress = { value: 0 };
-  let newprogress = 0;
-  let prevprogress = 0;
 
   // message
   const chats = document.getElementById('chats');
@@ -62,79 +56,29 @@ $(() => {
     return $('<div/>').text(input).text();
   }
 
-  function onReceiveMessageCallback(event) {
-    receiveBuffer.push(event.data);
-    receivedSize += event.data.byteLength;
-    receivedProgress.value = receivedSize;
-    newprogress = (receivedProgress.value / fileRec.size) * 100;
-    $('#file1').attr('aria-valuenow', newprogress).css('width', `${newprogress}%`);
-    $('#fileProgress').text(`Progress- ${Math.round(newprogress)}%`);
-    if (newprogress > prevprogress + 1) {
-      prevprogress = newprogress;
-      socket.emit('received-chunks', {
-        username: ExchangerUsername,
-        progress: newprogress,
-      });
-    }
-    if (receivedSize === fileRec.size) {
-      console.log('RECEIVED ENTIRE FILE');
-      socket.emit('received-chunks', {
-        username: ExchangerUsername,
-        progress: 100,
-      });
-      $('#fileProgress').text('Progress- 100%');
-      const received = new window.Blob(receiveBuffer);
-      receiveBuffer = [];
-      console.log('converted array to blob');
-      downloadAnchor.href = URL.createObjectURL(received);
-      console.log(downloadAnchor.href);
-      downloadAnchor.download = fileRec.name;
-      $progressBar.fadeIn();
-      $('#download').show();
-      $('#file-send-button').prop('disabled', false);
-      sender = false;
-      receiveBuffer = [];
-      receivedSize = 0;
-      receivedProgress = {
-        value: 0,
-      };
-      sendProgress = {
-        value: 0,
-      };
-      newprogress = 0;
-      prevprogress = 0;
-      $('#download').trigger('click');
-      console.log('Download link created');
-    }
-  }
-
   function cancelConnection() {
     offersForMe = [];
     socket.partner = null;
     socket.partnerid = null;
 
-    console.log('Connection terminated');
+    //console.log('Connection terminated');
     $transferPage.fadeOut();
     $progressBar.fadeOut();
     $homePage.show();
-    myPeerConn.close();
-    dataChannel.close();
+    $transferPageHeader.html('');
+    $downloadAnchor.fadeOut();
+    $downloadAnchor.prop('href','');
+    $downloadAnchor.html('');
+    // Clear the requests
+    $('.request-list').html('');
+
     ExchangerUsername = null;
     // offerComplete = false;
     sender = false;
-    receiveBuffer = [];
-    receivedSize = 0;
-    receivedProgress = {
-      value: 0,
-    };
-    sendProgress = {
-      value: 0,
-    };
-    newprogress = 0;
-    prevprogress = 0;
-    client.destroy();
-    // Clear the requests
-    $('.request-list').html('');
+    if (client !== null) {
+      client.destroy();
+      client = null;
+    }
   }
 
   function requestHandler(answer, btn) {
@@ -142,13 +86,12 @@ $(() => {
     if (answer === 'y') {
       socket.partner = requestingUsername;
       socket.partnerid = offersForMe[requestingUsername];
-      console.log(`${socket.partner} ${socket.partnerid}`);
+      //console.log(`${socket.partner} ${socket.partnerid}`);
       //    if request accepted
       ExchangerUsername = requestingUsername;
       // Set data-channel response on the other end(the client who receives the offer)
       $homePage.hide();
       $transferPage.fadeIn();
-      const $transferPageHeader = $('.user-name');
       $transferPageHeader.html(`<p>You are now connected to ${socket.partner}. To go back click <a href="#" class="alert-link" id="backLink"> here </a>. </p>`);
     } else {
       //    if request rejected
@@ -162,29 +105,27 @@ $(() => {
     // emit an event to the requesting user
   }
 
-  function sendData() {
-    $('#fileBeingSent').text(file.name + "(" + Math.round(file.size / 1000) + " KB)");
-
-    getClient();
-    var torrent = client.seed(file, function (torrent) {});
-
-    torrent.on("infoHash",function(){
-      console.log(torrent.magnetURI);
-      socket.emit("send", {
-        user:ExchangerUsername,
-        hash:torrent.magnetURI
-      });
-    });
-  }
-
   function getClient() {
     client = new WebTorrent({
-      tracker:{
-        rtcConfig: configuration}
+      tracker: { rtcConfig: configuration },
     });
+    client.on('error', (err) => { alert(err); });
+    client.on('warning', (w) => { alert(w); });
+  }
 
-    client.on("error",function(err){alert(err)});
-    client.on("warning",function(w){alert(w)})
+  function sendData() {
+    $('#fileBeingSent').text(`${file.name} ( ${Math.round(file.size / 1000)}  KB)`);
+
+    getClient();
+    const torrent = client.seed(file, () => {});
+
+    torrent.on('infoHash', () => {
+      //console.log(torrent.magnetURI);
+      socket.emit('send', {
+        user: ExchangerUsername,
+        hash: torrent.magnetURI,
+      });
+    });
   }
 
   $('#download').hide();
@@ -230,29 +171,35 @@ $(() => {
     $('#file-desc').text(file.name);
   });
 
-  $('#file-send-button').click(() => {
-    console.log(`${username} sending message`);
+  fileSendButton.click(() => {
+    //console.log(`${username} sending message`);
     const input = document.getElementById('file-1');
     if (!input) {
-      console.log("Um, couldn't find the fileinput element.");
+      alert("Um, couldn't find the fileinput element.");
       return;
     } else if (!input.files) {
-      console.log("This browser doesn't seem to support the `files` property of file inputs.");
+      alert("This browser doesn't seem to support the `files` property of file inputs.");
       return;
     } else if (!input.files[0]) {
-      console.log("Please select a file before clicking 'Load'");
+      alert("Please select a file before clicking 'Load'");
       return;
     }
     [file] = input.files;
 
-    console.log(`File is ${[file.name, file.size, file.type, file.lastModifiedDate].join(' ')}`);
+    //console.log(`File is ${[file.name, file.size, file.type, file.lastModifiedDate].join(' ')}`);
+
+    const fileStatus = `<li class = 'chatbox-file-history-sent'>  Sending  ${file.name} to ${ExchangerUsername}. </li>`;
+    $(fileStatus).appendTo($chatbox);
+
     statusMessage.textContent = '';
-    downloadAnchor.textContent = '';
-    if (file == null) console.log('No file selected');
+    $('#file1').attr('aria-valuenow', 0).css('width', '0%');
+    $downloadAnchor.fadeOut();
+
+    if (file == null) alert('No file selected');
     if (file.size === 0) {
       bitrateDiv.innerHTML = '';
       statusMessage.textContent = 'File is empty, please select a non-empty file';
-      console.log('File is empty, please select a non-empty file');
+      alert('File is empty, please select a non-empty file');
       // closeDataChannels();
       return;
     }
@@ -283,7 +230,7 @@ $(() => {
 
   $(document).on('click', '.user-name a', () => {
     //   cancel the for both users
-    console.log('Connection terminated');
+    //console.log('Connection terminated');
     socket.emit('Cancel Connection', ExchangerUsername);
     cancelConnection();
   });
@@ -298,12 +245,17 @@ $(() => {
     requestHandler('n', $(this));
   });
 
-  $("#stop-progress").click(()=>{
+  $('#stop-progress').click(() => {
     client.destroy();
+    client = null;
     sender = false;
-    $('#file-send-button').prop('disabled', false);
-    $("#fileProgress").text("Cancelled");
-    socket.emit("reject",ExchangerUsername);
+    fileSendButton.prop('disabled', false);
+
+    socket.emit('reject', ExchangerUsername);
+    $('#fileProgress').text('Cancelled');
+    
+    const fileStatus = `<li class = 'chatbox-file-history-cancel'>  You cancelled file transfer. </li>`;
+    $(fileStatus).appendTo($chatbox);
   });
 
   // messages
@@ -337,7 +289,7 @@ $(() => {
 
 
   socket.on('offer', (data) => {
-    console.log(`My Username is ${username}`);
+    //console.log(`My Username is ${username}`);
 
     offersForMe[data.username] = data.pid;
     // show that username wants to connect to you
@@ -356,9 +308,7 @@ $(() => {
   socket.on('updateUsersList', (onlineUsers) => {
     let html = '';
     for (let i = 0; i < onlineUsers.length; i += 1) {
-      if (onlineUsers[i] === username) {
-        continue;
-      } else {
+      if (onlineUsers[i] !== username) {
         html += `<div class="user"><button type="button" class="btn btn-default btn-block online-user" data-toggle="modal" data-target="#waiting_message">${onlineUsers[i]}</button> </div>`;
       }
     }
@@ -373,15 +323,12 @@ $(() => {
       socket.partner = msg.partner;
       socket.partnerid = msg.partnerid;
 
+      socket.emit('ack', msg);
       // stop the progress loader
       $homePage.hide();
       $transferPage.fadeIn();
       const $transferPageHeader = $('.user-name');
-      $transferPageHeader.html(`<p>You are now connected to ${socket.partner}${$transferPageHeader.html()}</p>`);
-      start(); // start the peerconnection process
-      // create peer connection offer and send
-      //   local description on other side(also create a data channel)
-      sendLocalDesc();
+      $transferPageHeader.html(`<p>You are now connected to ${socket.partner}. To go back click <a href="#" class="alert-link" id="backLink"> here </a>. </p>`);
     } else {
       // remove modal after informing partner has said no
       ExchangerUsername = null; // else set ExchangeUsername to None
@@ -389,32 +336,33 @@ $(() => {
   });
 
   socket.on('file-desc', (fileDesc) => {
-    console.log('file-desc received');
+    //console.log('file-desc received');
     if (!sender) { // to make sure we have not already sent a file offer to the other client
       // (we will disable the send button on other side to make sure only one person
       // sends a file at a time, but just to be sure)
-      console.log(`File is ${[fileDesc.name, fileDesc.size, fileDesc.type, fileDesc.lastModifiedDate].join(' ')}`);
+      //console.log(`File is ${[fileDesc.name, fileDesc.size, fileDesc.type, fileDesc.lastModifiedDate].join(' ')}`);
       fileRec = {
         name: fileDesc.name,
         size: fileDesc.size,
         type: fileDesc.type,
         lastModifiedDate: fileDesc.lastModifiedDate,
       };
-      console.log(`file offer of ${fileRec.name} accepted`);
+      //console.log(`file offer of ${fileRec.name} accepted`);
       socket.emit('file accepted', {
         target: ExchangerUsername,
         from: username,
         file: fileDesc.name,
       }); // can put a feature later to ask the user whether
       // he/she wants to accept the file, and based on that respond as accepted/refused
-      $('#file-send-button').prop('disabled', true);
+      fileSendButton.prop('disabled', true);
+      $('#file1').attr('aria-valuenow', 0).css('width','0%');
       $('#fileBeingSent').text(`${fileRec.name}(${Math.round(fileRec.size / 1000)} KB) (receiving..)`);
     } else {
       sender = false; // if both have sent at the same time, cancel both
-      console.log('file refused');
+      //console.log('file refused');
       socket.emit('file refused', ExchangerUsername);
     }
-    console.log('end');
+    //console.log('end');
   });
 
   socket.on('file refused', () => {
@@ -424,97 +372,101 @@ $(() => {
 
   socket.on('file accepted', (data) => { // This is for sender's end. Here funtion gets the username of the user he will now send the file to
     // here's the sendData!
-    console.log('trying to send');
+    //console.log('trying to send');
+
     $progressBar.fadeIn();
+    $('#stop-progress').html('Cancel');
+    $('#fileProgress').text('Establishing Connection');
+
     sendData(); // start sending :)))
-    console.log('send completed');
-    socket.emit('status', data);// want to tell the user that has sent the file that file has been sent. Here can add more info to put in file shar history
-    $progressBar.fadeOut();// this works (this is for sender's side)
-    const filehistory = `<li class = 'chatbox-file-history-sent'>  You sent ${data.file} to ${data.target}.  </li>`;
-    $(filehistory).prependTo($chatbox);// delivering file history to chat box of the sender
-  });
+    //console.log('send completed');
+    });
 
-  socket.on('status', (data) => { // this should fade out the progress bar after file seding is complete (this is only for reciver side)
-    $progressBar.fadeOut();
-    console.log(`${data.target}sent${data.file}to${data.from}`);
-    // class of the chat/file share history ul is chat
-    const filehistory = `<li class = 'chatbox-file-history-recieved'>  You recieved  ${data.file} from ${data.from}. </li>`;
-    $(filehistory).prependTo($chatbox);// delivering file history to chat box
-  });
-
-  socket.on('received-chunks', (prog) => {
-    newprogress = prog;
-    $('#file1').attr('aria-valuenow', newprogress).css('width', `${newprogress}%`);
-    $('#fileProgress').text(`Progress- ${Math.round(newprogress)}%`);
-    if (newprogress === 100) {
-      $('#file-send-button').prop('disabled', false);
-      sender = false;
-      receiveBuffer = [];
-      receivedSize = 0;
-      receivedProgress = {
-        value: 0,
-      };
-      sendProgress = {
-        value: 0,
-      };
-      newprogress = 0;
-      prevprogress = 0;
-    }
-  });
-
-  socket.on("send", function (hash){
+  socket.on('send', (hash) => {
     getClient();
-    console.log(hash);
-    client.add(hash,function(torrent){
-      console.log(hash);
-      torrent.on('error',(err)=>alert(err));
+    //console.log(hash);
 
-      torrent.on('download',function () {
+    const fileStatus = `<li class = 'chatbox-file-history-recieved'>  Receiving  ${fileRec.name} from ${ExchangerUsername}. </li>`;
+    $(fileStatus).appendTo($chatbox);
+
+    $progressBar.fadeIn();
+      $('#stop-progress').html('Cancel');
+      $('#fileProgress').text('Establishing Connection');
+    
+    client.add(hash, (torrent) => {
+
+      [file] = torrent.files;
+
+      torrent.on('error', (err) => { alert(err); });
+
+      torrent.on('download', () => {
         $progressBar.fadeIn();
-        var progress = torrent.progress *100;
-        $('#file1').attr('aria-valuenow', progress).css('width', progress  + '%');
-        $("#fileProgress").text("Progress- "+ Math.round(progress) + "%");
+        const progress = torrent.progress * 100;
+        $('#file1').attr('aria-valuenow', progress).css('width', `${progress}%`);
+        $('#fileProgress').text(`Progress- ${Math.round(progress)}%`);
 
-        socket.emit("progress",{
-          user:ExchangerUsername,
-          progress: progress
+        socket.emit('progress', {
+          user: ExchangerUsername,
+          prog: progress,
         });
-        if (progress==100){client.destroy}
       });
 
-      torrent.on('done',function(){
-        var file = torrent.files[0];
+      torrent.on('done', () => {
+        [file] = torrent.files;
 
-        file.getBlobURL(function(error,url){
-          if(error) return;
-          downloadAnchor.href = url;
-          downloadAnchor.download =file.name
-          downloadAnchor.textContent = "Download"+file.name;
-          $("#download").show();
+        $progressBar.fadeOut();
+
+        // class of the chat/file share history ul is chat
+        const filehistory = `<li class = 'chatbox-file-history-recieved'>  You recieved  ${file.name} from ${ExchangerUsername}. </li>`;
+        $(filehistory).appendTo($chatbox);// delivering file history to chat box
+        
+        file.getBlobURL((error, url) => {
+          if (error) {alert(error);return};
+          //console.log("file is here");
+          $downloadAnchor.prop('href', url);
+          $downloadAnchor.prop('download', file.name);
+          $downloadAnchor.text(`Download ${file.name}`);
+          $('#download').show();
+
+          client.destroy();
+          client = null;
+          fileSendButton.prop('disabled', false);
         });
       });
     });
   });
 
-  socket.on("progress",function(progress){
-    $progressBar.fadeIn();
-    $("#stop-progress").html("Cancel");
-    $('#file1').attr('aria-valuenow', progress).css('width', progress  + '%');
-    $("#fileProgress").text("Progress- "+ Math.round(progress) + "%");
-    if (progress==100){client.destroy()}
+  socket.on('progress', (progress) => {
 
+    $('#file1').attr('aria-valuenow', progress).css('width', `${progress}%`);
+    $('#fileProgress').text(`Progress- ${Math.round(progress)}%`);
+    if (progress === 100) { 
+      client.destroy(); 
+      client = null;
+      sender = false;
+
+      $progressBar.fadeOut();
+      const filehistory = `<li class = 'chatbox-file-history-sent'>  You sent ${file.name} to ${ExchangerUsername}.  </li>`;
+      $(filehistory).appendTo($chatbox);// delivering file history to chat box of the sender
+    }
   });
 
-  socket.on("reject",function(){
+  socket.on('reject', () => {
     client.destroy();
-    $("#fileProgress").text("Cancelled");
-    sender=false;
-    $('#file-send-button').prop('disabled', false);
+    client = null;
+
+    $('#fileProgress').text('Cancelled');
+
+    const fileStatus = `<li class = 'chatbox-file-history-cancel'>  Transfer cancelled by ${ExchangerUsername}. </li>`;
+    $(fileStatus).appendTo($chatbox);
+    
+    sender = false;
+    fileSendButton.prop('disabled', false);
   });
 
   socket.on('PartnerDisconnected', () => {
     // stop transfer or show dialog that partner has been disconnected retry from main page
-    console.log('Partner disconnected');
+    //console.log('Partner disconnected');
     alert('Your partner has disconnected');
     cancelConnection();
   });
@@ -524,7 +476,7 @@ $(() => {
   });
 
   socket.on('cancel', (dat) => {
-    console.log(dat);
+    //console.log(dat);
     let html = '';
     const $requestList = $('.request-list');
     for (let i = 0; i < dat.length; i += 1) {
@@ -554,4 +506,5 @@ $(() => {
     }
   }, 60000);
 });
+
 },{}]},{},[1]);
